[75] Add [[nodiscard]], [[unlikely]], parenthesize (std::min), use fixed-width integers
Enforce C++23 quality attributes across the codebase:

- [[nodiscard]] on 7 value-returning functions (SymbolKey::operator==,
  hash/equal, makeKey/makeKeyFromBuffer, onMarketEvent, try_pop)
- [[unlikely]] on 5 error/guard branches in RiskManager (null guard,
  position limit, notional limit) and SimpleMarketMakingStrategy
  (non-trade filter, sub-tick price guard)
- Parenthesize (std::min) in PositionManager for MSVC macro compat
- Replace size_t with uint32_t in pin_thread_to_core and loop index
- Fix test that discarded try_pop return value

+++ include/TradingEngine.hpp
@@ -7,6 +7,7 @@
 #include "PositionManager.hpp"
 #include "SimpleMarketMakingStrategy.hpp"
 #include <atomic>
+#include <cstdint>
 #include <memory>
 #include <string>
 #include <thread>
@@ -14,7 +15,7 @@
 
 class IFillListener;
 
-void pin_thread_to_core(std::thread &t, size_t core_id);
+void pin_thread_to_core(std::thread &t, uint32_t core_id);
 
 class TradingEngine {
 public:
